Finalizando otimização (tipos de dados)

+++ scripts/ddl/tabelas/DDL-geral.sql
@@ -20,7 +20,7 @@ go
 
 -- Tabela status_seguros
 create table status_seguros(
-	cd_status_seguros int not null primary key,
+	cd_status_seguros tinyint not null primary key,
 	status varchar(20) not null check(status in ('Ativo','Inativo'))
 )
 go
@@ -30,7 +30,7 @@ go
 create table seguros(
 	cd_seguro int not null primary key,
 	nm_seguradora varchar(50) not null,
-	cd_status_seguros int not null,
+	cd_status_seguros tinyint not null,
 	constraint fk_status_seguros foreign key (cd_status_seguros)
 	references status_seguros(cd_status_seguros)
 )
@@ -43,7 +43,7 @@ create table veiculos(
 	cd_cliente int not null,
 	marca varchar(20) not null,
 	modelo varchar(20) not null,
-	ano int not null,
+	ano smallint not null,
 	cd_seguro int not null,
 	constraint fk_cliente foreign key (cd_cliente)
 	references clientes(cd_cliente),
@@ -55,7 +55,7 @@ go
 
 -- Tabela status_manutencoes
 create table status_manutencoes(
-	cd_status_manutencoes smallint not null primary key,
+	cd_status_manutencoes tinyint not null primary key,
 	status varchar(40) not null check(status in ('Pronto', 'Em Andamento', 'Atrasado', 'Em espera de peças', 'Em espera de pagamento', 'Esperando o Seguro', 'Carro não veio', 'Em segundo Plano', 'Emergência', 'Sem espaço para armazenar'))
 )
 go
@@ -63,7 +63,7 @@ go
 
 -- Tabela alas
 create table alas(
-	cd_alas smallint not null primary key,
+	cd_alas tinyint not null primary key,
 	ala varchar(50) not null,
 	descricao varchar(200) null
 )
@@ -72,7 +72,7 @@ go
 
 -- Tabela status_agendamentos
 create table status_agendamentos(
-	cd_status_agendamento int not null primary key,
+	cd_status_agendamento tinyint not null primary key,
 	status varchar(20) not null check(status in ('Ativo','Cancelado','Retificada'))
 )
 go
@@ -80,7 +80,7 @@ go
 
 -- Tabela status_funcionarios
 create table status_funcionarios(
-	cd_status_funcionarios int not null primary key,
+	cd_status_funcionarios tinyint not null primary key,
 	status varchar(10) not null check(status in ('Ativo','Suspenso','Demitido','Férias'))
 )
 go
@@ -96,7 +96,7 @@ create table funcionarios(
 	data_entrada date not null,
 	data_saida date null,
 	sexo char(1) not null check(sexo in ('F','M')),
-	cd_status_funcionario int not null,
+	cd_status_funcionario tinyint not null,
 	constraint fk_status_funcionario foreign key (cd_status_funcionario)
 	references status_funcionarios(cd_status_funcionarios) 
 )
@@ -105,7 +105,7 @@ go
 
 -- Tabela tipos_manutencao
 create table tipos_manutencao(
-	cd_tipo int not null primary key,
+	cd_tipo smallint not null primary key,
 	descricao varchar(50) not null
 )
 go
@@ -119,8 +119,8 @@ create table agendamentos(
 	horario_agendado time not null,
 	cd_funcionario	int not null,
 	placa varchar(7) not null,
-	cd_tipo int not null,
-	cd_status_agendamento int not null,	
+	cd_tipo smallint not null,
+	cd_status_agendamento tinyint not null,	
 	constraint fk_status_agendamento foreign key (cd_status_agendamento)
 	references status_agendamentos(cd_status_agendamento),
 	constraint fk_tipo_agendamento foreign key (cd_tipo)
@@ -135,13 +135,13 @@ go
 
 -- Tabela manutencoes
 create table manutencoes(
-	cd_manutencao smallint not null primary key,
+	cd_manutencao int not null primary key,
 	placa varchar(7) not null,
-	cd_funcionario smallint not null,
+	cd_funcionario int not null,
 	cd_tipo smallint not null,
-	cd_alas smallint not null,
+	cd_alas tinyint not null,
 	vl_manutencao numeric(5,2) null,
-	cd_status_manutencoes smallint not null,
+	cd_status_manutencoes tinyint not null,
 	constraint fk_status_manutencao foreign key (cd_status_manutencoes)
 	references status_manutencoes(cd_status_manutencoes),
 	constraint fk_alas foreign key (cd_alas)
